refactor: page.tsx를 server side로 변경하고 페이지명을 추가하기 (#162)
* refactor: /application 페이지 전환

* refactor: /application/apply 페이지 전환

* refactor: /community/[boardCode] 페이지 전환

* refactor: /community/[boardCode]/[postId] 페이지 전환

* refactor: /community/[boardCode]/[postId]/modify 페이지 전환

* refactor: /login 페이지 전환

* refactor: /my 페이지 전환

* refactor: /my/favorite 페이지 전환

* refactor: /my/modify 페이지 전환
