+++ src/app/community/[boardCode]/[postId]/page.tsx
@@ -1,86 +1,25 @@
-"use client";
+import { Metadata } from "next";
 
-import { useRouter } from "next/navigation";
-import { useEffect, useState } from "react";
+import PostPageContent from "./PostPageContent";
 
-import { getPostDetailApi } from "@/services/community";
+interface PostPageProps {
+  params: {
+    boardCode: string;
+    postId: string;
+  };
+}
 
-import TopDetailNavigation from "@/components/layout/TopDetailNavigation";
-import CloudSpinnerPage from "@/components/loading/CloudSpinnerPage";
+export const metadata: Metadata = {
+  title: "게시글",
+};
 
-import CommentWrite from "./CommentWrite";
-import Comments from "./Comments";
-import KebabMenu from "./KebabMenu";
-import Post from "./Post";
-
-import { Post as PostType } from "@/types/community";
-
-const PostPage = ({ params }: { params: { boardCode: string; postId: string } }) => {
-  const router = useRouter();
+const PostPage = ({ params }: PostPageProps) => {
   const { boardCode } = params;
   const postId = Number(params.postId);
-  const [post, setPost] = useState<PostType | null>(null);
-  const [isLoading, setIsLoading] = useState<boolean>(true);
-  const [curSelectedComment, setCurSelectedComment] = useState<number | null>(null);
-
-  const fetchPosts = async () => {
-    await getPostDetailApi(postId)
-      .then((res) => {
-        setPost(res.data);
-      })
-      .catch((err) => {
-        if (err.response) {
-          console.error("Axios response error", err.response);
-          if (err.response.status === 401 || err.response.status === 403) {
-            alert("로그인이 필요합니다");
-            document.location.href = "/login";
-          } else {
-            alert(err.response.data?.message);
-          }
-        } else {
-          console.error("Error", err.message);
-          alert(err.message);
-        }
-      });
-    setIsLoading(false);
-  };
-
-  useEffect(() => {
-    fetchPosts();
-  }, [postId]);
-
-  if (isLoading || post === null) {
-    return <CloudSpinnerPage />;
-  }
 
   return (
     <>
-      {/* 페이지 타이틀을 설정하려면 metadata API를 사용할 수 있습니다 */}
-      <TopDetailNavigation
-        title={post.postFindBoardResponse.koreanName}
-        handleBack={() => {
-          router.push(`/community/${boardCode}`);
-        }}
-        icon={post.isOwner && <KebabMenu boardCode={boardCode} postId={postId} />}
-      />
-      <div>
-        <Post post={post} boardCode={boardCode} postId={postId} />
-        <Comments
-          comments={post.postFindCommentResponses}
-          postId={postId}
-          refresh={() => {
-            router.refresh();
-          }}
-          setCurSelectedComment={setCurSelectedComment}
-          onSuccess={fetchPosts}
-        />
-        <CommentWrite
-          postId={postId}
-          curSelectedComment={curSelectedComment}
-          setCurSelectedComment={setCurSelectedComment}
-          onSuccess={fetchPosts}
-        />
-      </div>
+      <PostPageContent boardCode={boardCode} postId={postId} />
     </>
   );
 };
