+++ src/app/my/favorite/page.tsx
@@ -1,76 +1,20 @@
-"use client";
+import { Metadata } from "next";
 
-import Head from "next/head";
-import { useEffect, useState } from "react";
-
-import { getMyWishUniversityApi } from "@/services/myInfo";
-
-import UniversityCards from "@/components/college/UniversityCards";
 import TopDetailNavigation from "@/components/layout/TopDetailNavigation";
-import ScrollTab from "@/components/ui/ScrollTab";
 
-import { ListUniversity } from "@/types/university";
+import FavoriteContent from "./FavoriteContent";
 
-const MyScrapPage = () => {
-  const [wishColleges, setWishColleges] = useState<ListUniversity[] | null>(null);
+export const metadata: Metadata = {
+  title: "즐겨찾기",
+};
 
-  useEffect(() => {
-    const fetchWishColleges = async () => {
-      await getMyWishUniversityApi()
-        .then((res) => {
-          setWishColleges(res.data);
-        })
-        .catch((err) => {
-          if (err.response) {
-            console.error("Axios response error", err.response);
-            if (err.response.status === 401 || err.response.status === 403) {
-              alert("로그인이 필요합니다");
-              document.location.href = "/login";
-            } else {
-              alert(err.response.data?.message);
-            }
-          } else {
-            console.error("Error", err.message);
-            alert(err.message);
-          }
-        });
-    };
-    fetchWishColleges();
-  }, []);
-
-  // const tabs = ["스크랩 한 글", "멘토", "위시학교"];
-  const tabs = ["위시학교"];
-  const [tab, setTab] = useState(tabs[0]);
-
-  if (!wishColleges) {
-    return null;
-  }
-
+const FavoritePage = () => {
   return (
     <>
-      <Head>
-        <title>즐거찾기</title>
-      </Head>
-      <TopDetailNavigation title="즐거찾기" />
-      <ScrollTab choices={tabs} choice={tab} setChoice={setTab} />
-      {(() => {
-        switch (tab) {
-          case tabs[0]:
-            return (
-              <div className="mt-5">
-                <UniversityCards colleges={wishColleges} />
-              </div>
-            );
-          // case tabs[1]:
-          // return <PostCards posts={posts} />;
-          // case tabs[2]:
-          // return <></>;
-          default:
-            return null;
-        }
-      })()}
+      <TopDetailNavigation title="즐겨찾기" />
+      <FavoriteContent />
     </>
   );
 };
 
-export default MyScrapPage;
+export default FavoritePage;
